fix: :bug: Fix useBackend closure problem
`useBackend` take a factory function as parameter.
But the params of the call was not yet determine while in setup stage.
Those values are usually of type `Ref<T>`.

example:

```js
const username = ref("")
const password = ref("")
useBackend(logIn({username, password}), false)
```

`useBackend` is now modified to accept Ref type and non-Ref types.
It will wrap parameters in to a `Reactive`
and evaluate the internal value when it need to use.